requirements.txt
alembic==0.9.8
attrs==17.4.0
blinker==1.4
click==6.7
coverage==4.5.1
flake8==3.5.0
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
numpy==1.14.2
pandas==0.22.0
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytest==3.4.2
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.7.0
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
